requirements.txt
# Core LLM + API
openai>=1.30.0

# LangChain core (LCEL / Runnables)
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0

# Tavily search tool
tavily-python>=0.3.3

# Web scraping
beautifulsoup4>=4.12.3 
requests>=2.31.0   # tool for web scrapping 
lxml>=5.2.0

# Async support (important for multi-agent systems)
aiohttp>=3.9.5
httpx>=0.27.0

# Environment management
python-dotenv>=1.0.1

# Data handling / utilities
pydantic>=2.7.1
tqdm>=4.66.4

# Logging & debugging
loguru>=0.7.2

# Optional but useful for structured outputs
orjson>=3.10.3

# Tokenization (sometimes required by LangChain/OpenAI)
tiktoken>=0.7.0

# If you plan to expand to vector storage later
faiss-cpu>=1.8.0

# For better HTML parsing edge cases
html5lib>=1.1